requirements.txt
# Install the current package in editable mode
-e .

pydantic
pytest
websockets
python-dateutil
